Opdatering af diverse filer: bugfix, pænere grafer, navneændring, udvidelse af scripts
