+++ RTK_Statistik.py
@@ -202,42 +202,42 @@ plt.savefig("Figurer/Diffkronologisk_S_G_5D.png")
 # Histogram plot af differencer på hvert net og instrument
 smart_Leica.hist(column= 'Difference', bins=50)
 plt.axvline(mean_smart_Leica, color='w', linestyle='dashed', linewidth=2)
-plt.title('Leica på Smartnet. Gnst: ' + str(round(mean_smart_Leica,2)))
+plt.title('Leica på Smartnet \n Gnst: ' + str(round(mean_smart_Leica,2)) + 'mm')
 plt.ylabel("Antal")
 plt.xlabel("Difference [mm]")
 plt.savefig("Figurer/Histogram_H_H_5D.png")
 
 GPS_Leica.hist(column= 'Difference', bins=50)
 plt.axvline(mean_GPS_Leica, color='w', linestyle='dashed', linewidth=2)
-plt.title('Leica på GPSnet. Gnst: ' + str(round(mean_GPS_Leica,2)))
+plt.title('Leica på GPSnet \n Gnst: ' + str(round(mean_GPS_Leica,2)) + 'mm')
 plt.ylabel("Antal")
 plt.xlabel("Difference [mm]")
 plt.savefig("Figurer/Histogram_H_G_5D.png")
 
 smart_Trimble.hist(column= 'Difference', bins=50)
 plt.axvline(mean_smart_Trimble, color='w', linestyle='dashed', linewidth=2)
-plt.title('Trimble på Smartnet. Gnst: ' + str(round(mean_smart_Trimble,2)))
+plt.title('Trimble på Smartnet \n Gnst: ' + str(round(mean_smart_Trimble,2)) + 'mm')
 plt.ylabel("Antal")
 plt.xlabel("Difference [mm]")
 plt.savefig("Figurer/Histogram_G_H_5D.png")
 
 GPS_Trimble.hist(column= 'Difference', bins=50)
 plt.axvline(mean_GPS_Trimble, color='w', linestyle='dashed', linewidth=2)
-plt.title('Trimble på GPSnet. Gnst: ' + str(round(mean_GPS_Trimble,2)))
+plt.title('Trimble på GPSnet \n Gnst: ' + str(round(mean_GPS_Trimble,2)) + 'mm')
 plt.ylabel("Antal")
 plt.xlabel("Difference [mm]")
 plt.savefig("Figurer/Histogram_G_G_5D.png")
 
 smart_Sept.hist(column= 'Difference', bins=50)
 plt.axvline(mean_smart_Sept, color='w', linestyle='dashed', linewidth=2)
-plt.title('Septentrio på Smartnet. Gnst: ' + str(round(mean_smart_Sept,2)))
+plt.title('Septentrio på Smartnet \n Gnst: ' + str(round(mean_smart_Sept,2)) + 'mm')
 plt.ylabel("Antal")
 plt.xlabel("Difference [mm]")
 plt.savefig("Figurer/Histogram_S_H_5D.png")
 
 GPS_Sept.hist(column= 'Difference', bins=50)
 plt.axvline(mean_GPS_Sept, color='w', linestyle='dashed', linewidth=2)
-plt.title('Septentrio på GPSnet. Gnst: ' + str(round(mean_GPS_Sept,2)))
+plt.title('Septentrio på GPSnet \n Gnst: ' + str(round(mean_GPS_Sept,2)) + 'mm')
 plt.ylabel("Antal")
 plt.xlabel("Difference [mm]")
 plt.savefig("Figurer/Histogram_S_G_5D.png")
